no more showing figures in table to save time, delete stucked experiments, add config file to cache db path

+++ pyerm/webUI/tables.py
@@ -20,7 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-# Version: 0.2.6
+# Version: 0.2.7
 
 import pandas as pd
 from PIL import Image
@@ -29,10 +29,11 @@ from io import BytesIO
 import streamlit as st
 import os
 import re
-import tempfile
+from time import strftime, gmtime
 
 from pyerm.database.utils import delete_failed_experiments
 from pyerm.database.dbbase import Database
+from pyerm.webUI import PYERM_HOME
 
 def tables():
     title()
@@ -52,23 +53,27 @@ def detect_tables():
     st.session_state.table_name = table_name
 
 def select_tables():
-    def image_to_base64(img):
-        buffered = BytesIO(img)
-        img_str = base64.b64encode(buffered.getvalue()).decode()
-        return img_str
+    # def image_to_base64(img, desc):
+    #     buffered = BytesIO(img)
+    #     img_str = base64.b64encode(buffered.getvalue()).decode()
+    #     return img_str
     
-    def make_image_clickable(image_name, image):
-        img_str = image_to_base64(image)
-        return f'<a href="data:image/jpeg;base64,{img_str}" target="_blank" title="{image_name}"><img src="data:image/jpeg;base64,{img_str}" width="100"></a>'
-
+    # def make_image_clickable(image_name, image, desc):
+    #     img_str = image_to_base64(image, desc)
+    #     if img_str:
+    #         return f'<a href="data:image/jpeg;base64,{img_str}" target="_blank" title="{image_name}"><img src="data:image/jpeg;base64,{img_str}" width="100"></a>'
+    #     else:
+    #         return f'<a href="#" title="None"><img src="#" width="100"></a>'
+    
     def fold_detail_row(row, col_name):
         if row[col_name]:
-            return f'<details><summary>Details</summary>{row[col_name]}</details>'
+            detail = row[col_name].replace("\n", "<br>")
+            return f'<details><summary>Details</summary>{detail}</details>'
         else:
             return 'None'
 
     db = Database(st.session_state.db_path, output_info=False)
-    table_name = st.session_state.table_name
+    table_name:str = st.session_state.table_name
     if st.session_state.sql is not None:
         try:
             table_name = st.session_state.table_name
@@ -83,37 +88,40 @@ def select_tables():
         columns = [column[0] for column in db.cursor.description]
         df = pd.DataFrame(data, columns=columns)
     
-    # special process for image columns
-    columns_keep = [col for col in df.columns if not col.startswith("image_")]
-    pattern = re.compile(r'image_(\d+)')
-    max_image_num = -1
-    for name in df.columns:
-        match = pattern.match(name)
-        if match:
-            max_image_num = max(max_image_num, int(match.group(1)))
-    for i in range(max_image_num+1):
-        if f'image_{i}' in df.columns and not df[f'image_{i}_name'].isnull().all():
-            df[f'image_{i}'] = df.apply(lambda x: make_image_clickable(x[f'image_{i}_name'], x[f'image_{i}']), axis=1)
-            columns_keep.append(f'image_{i}')
-            
-    if "failed_reason" in df.columns:
-        df['failed_reason'] = df.apply(lambda x: fold_detail_row(x, 'failed_reason'), axis=1)
-    df = df[columns_keep]
     
-    if st.button('Refresh'):
+    if st.button('Refresh', key='refresh1'):
         st.session_state.table_name = table_name
     st.write('## Table:', table_name)
+    
+    columns_keep = [col for col in df.columns if not col.startswith("image_")]
+    
     if table_name == 'experiment_list':
-        if st.checkbox('Delete all failed records'):
+        if st.checkbox('Delete all failed and stuck records'):
             st.write('**Warning: This operation will delete all failed records and their results, which cannot be undone.**')
             if st.button(f"Confirm"):
                 delete_failed_experiments(db)
                 st.session_state.table_name = table_name
+        df['failed_reason'] = df.apply(lambda x: fold_detail_row(x, 'failed_reason'), axis=1)
+        df['useful_time_cost'] = df['useful_time_cost'].apply(lambda x: strftime('%H:%M:%S', gmtime(x)) if not pd.isnull(x) else x) 
+        df['total_time_cost'] = df['total_time_cost'].apply(lambda x: strftime('%H:%M:%S', gmtime(x)) if not pd.isnull(x) else x)
+    # elif table_name.startswith("result_"):
+    #     # special process for image columns
+    #     pattern = re.compile(r'image_(\d+)')
+    #     max_image_num = -1
+    #     for name in df.columns:
+    #         match = pattern.match(name)
+    #         if match:
+    #             max_image_num = max(max_image_num, int(match.group(1)))
+                
+    #     for i in range(max_image_num+1):
+    #         if f'image_{i}' in df.columns and not df[f'image_{i}_name'].isnull().all():
+    #             df[f'image_{i}'] = df.apply(lambda x: make_image_clickable(x[f'image_{i}_name'], x[f'image_{i}'], desc=f"{x[f'experiment_id']}_{i}"), axis=1)
+    #             columns_keep.append(f'image_{i}')
+    
+    df = df[columns_keep]
     st.write(df.to_html(escape=False, columns=columns_keep), unsafe_allow_html=True)
-    
-    # st.dataframe(df[columns_keep])
-
-    
+    if st.button('Refresh', key='refresh2'):
+        st.session_state.table_name = table_name
 
 def input_sql():
     st.sidebar.write('You can also set the columns and condition for construct a select SQL sentense for the current table here.')
